Restructure models file. Improve player addition and removal.
The classes in the models file are moved around to be more friendly
for type hints.

The method for converting `Player` to json is begun.

Rooms have a method that allows adding players by object instance
or uuid (primary key). This function must return an instance of the
object loaded from redis in order to represent any changes made.

Better comments that document json schema are written.

Signed-off-by: Tyler Jones <tdjones879@gmail.com>