Move walrus models to own module
Walrus/Redis models are moved to their own module to clean up import
problems. This will allow the project to be split into separate modules
for each 'service' that we want to provide.

`builtins.walrus_conn` must be set before importing the new module to
provide it with the correct connection to redis.