Move all web server components to separate package
All code concerned with socketIO events and http routing are moved to a
separate package. Will clean up some package import madness. This also
further allows us to have a separate package for each 'service' we are
providing.

The dockerfile is modified to call the new web module from the correct
directory to allow submodules to import models correctly. We should not
try to import from outside the web package except for inside
web/__init__.py. As such, if the web package needs to be run, navigate
to the parent `multipong` directory and call `python3 -u -m web` from
inside there.