Force docker-compose to create two containers
The same Dockerfile is used for both containers so that caching can
save time when bringing containers up and down. Containers are
differentiated by the command used when they are started.

Old containers must be removed from docker before use because images
were renamed and docker-compose will not recognize the change.