Subscribe to serverUpdate and send updates to clients
serverUpdate's are sent to client as testUpdate's. The backgroundThread
function is kept in web/sockets.py for now to avoid `Runtime Error:
working outside request context`.

Python modules are fubar and create copies of values when trying to
import socketio by calling `from . import socketio`, so we keep the
updates in the same module namespace as the rest of the socketio logic.